Bugfixes for duty cycle, particularly for subsampled Rx.  Also changed demo text to no longer have newlines

+++ src/mainThread.h
@@ -22,7 +22,7 @@ int32_t maxBlocksToProcess;
 int32_t fifoSize; //Size in blocks
 #endif
 double txDutyCycle;
-int rxSubsamplePeriod;
+int64_t rxSubsamplePeriod;
 TX_GAIN_DATATYPE gain;
 } threadArgs_t;
